Renamed file to auth_page.py
